Added services and repository for transfer

+++ src/Lykke.Service.PayInternal.Core/Domain/Transfer/ISourceAmount.cs
@@ -1,8 +1,17 @@
 ﻿namespace Lykke.Service.PayInternal.Core.Domain.Transfer
 {
+    /// <summary>
+    /// Source - amount pair
+    /// </summary>
     public interface ISourceAmount
     {
+        /// <summary>
+        /// Source Address
+        /// </summary>
         string SourceAddress { get; set; }
-        double Amount { get; set; }
+        /// <summary>
+        /// Amount
+        /// </summary>
+        decimal Amount { get; set; }
     }
 }
